MIM-38 拆分工作区加载回归测试
