+++ ios/MimiRemote/Tests/MimiRemoteTests/ConversationDirectRuntimeTests.swift
@@ -906,71 +906,6 @@ extension ConversationDataFlowTests {
         XCTAssertEqual(claudeCloseCount, 0)
     }
 
-    func testMultiRuntimeWorkspaceSessionListPropagatesClaudeFailureAfterCodexSuccess() async throws {
-        let project = AgentProject(id: "proj_multi_list_failure", name: "Multi List", path: "/tmp/multi-list")
-        let workspace = AgentWorkspace(project: project)
-        let config = makeDirectAppServerConfig(project: project, channels: [makeClaudeChannelMetadata()])
-        let codexTransport = FakeCodexAppServerTransport()
-        let claudeTransport = FakeCodexAppServerTransport()
-        let client = MultiRuntimeSessionAPIClient(
-            codexRuntime: CodexAppServerSessionRuntime(
-                endpoint: "http://127.0.0.1:8787",
-                token: "outer-token",
-                runtimeProvider: "codex",
-                transportFactory: { codexTransport },
-                configProvider: { config }
-            ),
-            claudeRuntime: CodexAppServerSessionRuntime(
-                endpoint: "http://127.0.0.1:8787",
-                token: "outer-token",
-                runtimeProvider: "claude",
-                transportFactory: { claudeTransport },
-                configProvider: { config }
-            )
-        )
-
-        let listTask = Task {
-            try await client.sessionsPage(
-                workspace: workspace,
-                cursor: nil,
-                limit: 20,
-                consistency: .authoritative
-            )
-        }
-        let codexInitialize = try await waitForFakeAppServerRequest(codexTransport, method: "initialize")
-        transportResponse(codexTransport, id: codexInitialize.id, result: #"{"userAgent":"fake-codex","platformFamily":"macos"}"#)
-        let codexList = try await waitForFakeAppServerRequest(codexTransport, method: "thread/list", after: 1)
-        transportResponse(
-            codexTransport,
-            id: codexList.id,
-            result: appServerThreadListResult([
-                appServerThreadJSON(
-                    id: "codex-visible-before-claude-failure",
-                    cwd: project.path,
-                    source: "appServer",
-                    updatedAt: 1_780_493_000
-                )
-            ], nextCursor: nil)
-        )
-
-        let claudeInitialize = try await waitForFakeAppServerRequest(claudeTransport, method: "initialize")
-        transportResponse(claudeTransport, id: claudeInitialize.id, result: #"{"userAgent":"fake-claude","platformFamily":"macos"}"#)
-        let claudeList = try await waitForFakeAppServerRequest(claudeTransport, method: "thread/list", after: 1)
-        transportErrorResponse(
-            claudeTransport,
-            id: claudeList.id,
-            code: -32000,
-            message: "Claude workspace list failed"
-        )
-
-        do {
-            _ = try await listTask.value
-            XCTFail("Claude 列表失败必须向 WorkspaceRootView 抛出，才能显示 failed 状态和重试入口")
-        } catch {
-            XCTAssertTrue(error.localizedDescription.contains("Claude workspace list failed"))
-        }
-    }
-
     func testMultiRuntimeSocketsKeepCodexAndClaudeConnectionsOnActiveHost() async throws {
         let project = AgentProject(id: "proj_multi_socket", name: "Multi Socket", path: "/tmp/multi-socket")
         let config = makeDirectAppServerConfig(project: project, channels: [makeClaudeChannelMetadata()])
